Preliminary tileset support.
OpenGL mode now supports graphical tiles. The implementation is
fairly rough, but is extensible, and should function nicely for
a long time. I'm using Kenney's 1-bit Pack as a placeholder,
since Kenney's Pack continues to be amazing.

I think the next step is to dynamically calculate UI size, as
opposed to defining it in the code. This should fix some of
the UI weirdness in the OpenGL port, and could also open up
the possibility of gracefully handling a sigwinch.