Re-enable winch handling.
Well, it's not exactly graceful, but the game doesn't hard crash
when the terminal is resized anymore. In theory, we probably should
add some checks so that users cannot segfault the game by resizing
to a very weird dimension, although I'm not sure whether or not
that is possible.

The OpenGL port does not handle resizing gracefully. Handling that
will be a much more involved process then handling a sigwinch.