Set up actor components.
This commit does a number of things:

First, the list of actors is now allocated on the heap, rather
than the stack. This was the intention from the beginning, but I
was lazy when just starting out and I didn't want to do the
necessary memory management.

Second, I set up the infrastructure necessary for a basic entity
component system by adding inventory and item components. Like the
AI component, they are effectively boilerplate code for now, but
in the next few commits that is going to change.

Finally, I fixed a memory leak in the window code that was exposed
when running the new system under valgrind.