Write global struct to save file.
Well this is embarrassing. Immediately after claiming that the
serialization code should not have to change very often, I
realized that I had not saved the global struct, meaning that
information such as turncount and depth were inconsistent across
reloads.

After loading the global struct, I ran into another problem, however.
The locale was written to the global save file, and there is a
small chance that a user might save a game on a terminal with one
locale, then load it on a terminal with a different one. To rectify
this, I moved the locale to the term struct, where it probably should
be anyway.

Saving and loading both function now, and both take place practically
instantly.