Parse actors from JSON.
This is a huge commit which does a lot. The biggest change here is
defining actors in JSON files, rather than directly in the code.
This should save us quite a bit of RAM, and also make modifying
the game much easier.

This commit also includes a lot of cleanup, such as changing all
instances of sprintf() to snprintf().