Action Dispatch Rewrite: Part One
I know that I need to work on actually adding content to the game,
but this is a nasty enough piece of refactoring that I would rather
take care of it immediately.

The goal of this refactoring is to allow for rebinding keys later in
the development process, to cut down on the number of massive switch
statements in the codebase, to make adding new windowports easier, and
to make adding new actions as easy as adding a single entry to an
array.